Create Car_arduino_xbee_debug
Everything in this code is commented out except for the xbee setup, ESC setup, and the Wheel servo setup. Even with this code the car was still jerking. This code never gives any commands to the ESC but every time a message is sent to the Xbee by the coordinator the wheels try to spin backward momentarily.